% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAQ.R
\name{CalcAQ}
\alias{CalcAQ}
\title{Calculate photosynthetic light response from Amax, phi, Rd and theta.}
\usage{
CalcAQ(Amax, phi, Rd, theta, Q = seq(from = 0, to = 2000, by = 50))
}
\arguments{
\item{Amax}{Maximum net assimilation rate}

\item{phi}{Quantum efficiency of the photosynthetic light response. I.e. initial slope at low light.}

\item{Rd}{Respiration rate (positive value!).}

\item{theta}{Curvature of the non-retangular hyperbola.}

\item{Q}{Vector of light levels (PPFD). By default \code{seq(from = 0, to = 2000, by = 50))} is used.}
}
\description{
This function calculates the light response of the net assimilation rate based on for parameters. Input paramters are maximum photosynthetic rate Amax, initial quantum efficiency phi, respiration rate Rd, and curvature theta. Using a non-rectangular hyperbola based on these four parameters, the net assimilation rate at a series of light levels is calcualted. This is the reverse of the function \code{FitAQ}.
}
\examples{
x <- CalcAQ(24, 0.07, 2.26, 0.7)
plot(A ~ Q, data = x)
}
\seealso{
\code{\link[FitAQ:FitAQ]{FitAQ}}
}
