% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_light_response.R
\name{fit_light_response}
\alias{fit_light_response}
\title{Fit light response curves using DEoptim}
\usage{
fit_light_response(
  data,
  Qin = NULL,
  A = NULL,
  model = c("rectangular", "nonrectangular", "exponential", "modified"),
  sat_frac = 0.9,
  type = c("LI-6800", "LI-6400", "other"),
  lower = NULL,
  upper = NULL,
  control = list(trace = FALSE),
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{Qin}{Name or index of PPFD column. If \code{NULL} and \code{type} is specified,
a default will be used.}

\item{A}{Name or index of net photosynthesis column. If \code{NULL} and \code{type} is specified,
a default will be used.}

\item{model}{Model type.}

\item{sat_frac}{Fraction of Pmax for LSP.}

\item{type}{Instrument type: \code{"LI-6800"}, \code{"LI-6400"}, or \code{"other"}.
If \code{"LI-6800"}, defaults: \code{Qin = "Qin"}, \code{A = "A"}.
If \code{"LI-6400"}, defaults: \code{Qin = "PARi"}, \code{A = "Photo"}.
If \code{"other"}, user must provide \code{Qin} and \code{A}.}

\item{lower, upper}{Optional bounds.}

\item{control}{List for DEoptim.}

\item{...}{Additional arguments.}
}
\description{
Fit light response curves using DEoptim
}
