% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitAQ.R
\name{FitAQ}
\alias{FitAQ}
\title{Flexible function to calculate a non-rectangular hyperbola fit for a photosynthetic light response curve.}
\usage{
FitAQ(data, A = A, Q = Q, nlscontrol = NULL, start = NULL,
  provide.model = FALSE, ...)
}
\arguments{
\item{data}{Data frame with the light response gas exchange data. Needs to provide vectors for light level and assimilation rate}

\item{A}{Name of the vector in the data frame that holds the assimilation rate.}

\item{Q}{Name of the vector in the data frame that holds the light level information.}

\item{nlscontrol}{Optional object that controls internal settings of nls. See ?nls.control. Default is NULL.}

\item{start}{Optional named list that provides starting values for nls fit of Amax, phi, Rd, theta. (Note: Rd is a positive value)}

\item{provide.model}{Logical. If FALSE (the default) the model coefficients are returned. When TRUE, the model is returned.}

\item{...}{Other options to the underlying functions.}
}
\value{
Depending on the value of provide.model, the data frame with the coefficients of a non-rectangular hyperbola (Amax, phi, Rd, theta), is returned. When provide.model = TRUE, the curve fit model is returned.
}
\description{
Fit a non-rectangular hyperbola to the photosynthetic light response (assimilation rate as a function of light) following Lambers et al. 2008, Thornley 1976, Farquhar & Sharkey 1982: \code{(A ~ ((phi * Q + Amax - sqrt((phi * Q + Amax)^2 - 4 * theta * phi * Q * Amax)) / 2 * theta) - Rd)} Uses \code{nls} for the curve fit using the algorithm \code{"port"}. Starting values for the initial curve fit (Amax, phi, Rd, theta) can be provided as a named list. Otherwise, a simple estimation of starting parameters takes place.
}
\references{
Farquhar, Graham D., and Thomas D. Sharkey. ‘Stomatal Conductance and Photosynthesis’. Annual Review of Plant Physiology 33, no. 1 (1982): 317–345.

Lambers, Hans, F. Stuart Chapin III, and Thijs L. Pons. Plant Physiological Ecology. Springer Science & Business Media, 2008.

Thornley, J. H. M. Mathematical Models in Plant Physiology : A Quantitative Approach to Problems in Plant and Crop Physiology. London; New York: Academic Press, 1976.
}
