% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLCP_FitSat.R
\name{FitLCP}
\alias{FitLCP}
\title{Calculate light compensation point and light saturation point from non-rectangular hyperbola model of a photosynthetic light reponse}
\usage{
FitLCP(model, coef.Amax = "Amax", coef.phi = "phi",
  coef.theta = "theta", coef.Rd = "Rd")
}
\arguments{
\item{model}{Model object from either \code{FitAQ} or \code{nls}}

\item{coef.Amax}{Character string with the name of the curve fit coefficient representing maximum photosynthetic assimilation rate Amax. Default is "Amax".}

\item{coef.phi}{Character string with the name of the curve fit coefficient representing the initial slope phi. Default is "phi".}

\item{coef.theta}{Character string with the name of the curve fit coefficient representing the curvature of the hyperbola. Default is "theta".}

\item{coef.Rd}{Character string with the name of the curve fit coefficient representing the respiration rate Rd. Default is "Rd".}
}
\value{
Light compensation point. Named numeric.
}
\description{
Calculate light compensation point and light saturation point from non-rectangular hyperbola model of a photosynthetic light reponse
}
