% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLCP_FitSat.R
\name{FitSat}
\alias{FitSat}
\title{Calculate the light level at which the photosynthetic assimilation is saturated}
\usage{
FitSat(model, sat.fac = 0.9, range = c(0, 2200), coef.Amax = "Amax",
  coef.phi = "phi", coef.theta = "theta", coef.Rd = "Rd")
}
\arguments{
\item{model}{Model object created by either \code{FitAQ} or \code{nls}.}

\item{sat.fac}{Degree of saturation of Amax. Defaults to 0.9, i.e. 90\% of Amax.}

\item{range}{Numeric list with two entries: First entry is the lower end of a window for the light level at which the assimilation rate is saturated. Second entry is the upper end of the window.}

\item{coef.Amax}{Character string with the name of the curve fit coefficient representing maximum photosynthetic assimilation rate Amax. Default is "Amax".}

\item{coef.phi}{Character string with the name of the curve fit coefficient representing the initial slope phi. Default is "phi".}

\item{coef.theta}{Character string with the name of the curve fit coefficient representing the curvature of the hyperbola. Default is "theta".}

\item{coef.Rd}{Character string with the name of the curve fit coefficient representing the respiration rate Rd. Default is "Rd".}
}
\value{
Light level at which the desired level of Amax is reached. Named numeric.
}
\description{
Given a odel of a non-recangular curve fit of assimilation rate and light, this funciton calculates the light level at which a specific proportion of Amax is reached.
}
