% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.light_response_fit}
\alias{print.light_response_fit}
\title{Print method for light_response_fit objects}
\usage{
\method{print}{light_response_fit}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{light_response_fit} object.}

\item{digits}{Number of digits to display.}

\item{...}{Not used.}
}
\description{
Print method for light_response_fit objects
}
