% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_light_response.R
\name{fit_light_response}
\alias{fit_light_response}
\title{Fit light response curves using DEoptim}
\usage{
fit_light_response(
  data,
  Qin,
  A,
  model = c("rectangular", "nonrectangular", "exponential", "modified"),
  sat_frac = 0.9,
  lower = NULL,
  upper = NULL,
  control = list(trace = FALSE),
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{Qin}{Name or index of PPFD column.}

\item{A}{Name or index of net photosynthesis column.}

\item{model}{Model type: "rectangular", "nonrectangular", "exponential", "modified".}

\item{sat_frac}{Fraction of Pmax to define saturation point (default 0.9).}

\item{lower, upper}{Optional named vectors of parameter bounds.}

\item{control}{List passed to \code{DEoptim.control}.}

\item{...}{Additional arguments to \code{DEoptim}.}
}
\value{
An object of class \code{light_response_fit}.
}
\description{
Fits one of four light response models using Differential Evolution optimization.
}
\examples{
\dontrun{
res <- fit_light_response(mydata, Qin = "PAR", A = "Photo", model = "exponential")
print(res)
plot(res)
light_points(res)
}
}
