% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_line.R
\name{low_light_coefs}
\alias{low_light_coefs}
\title{Linear regression on low-light points to estimate AQY and Rd}
\usage{
low_light_coefs(
  data,
  Qin = NULL,
  A = NULL,
  n = 5,
  type = c("LI-6800", "LI-6400", "other"),
  min_points = 3
)
}
\arguments{
\item{data}{Data frame.}

\item{Qin}{Name of light intensity column (optional if type given).}

\item{A}{Name of photosynthesis column (optional if type given).}

\item{n}{Number of lowest-light points (default 5).}

\item{type}{Instrument type: "LI-6800", "LI-6400", or "other". Default "LI-6800".}

\item{min_points}{Minimum points required.}
}
\value{
List with AQY, Rd, intercept, n_used, r_squared.
}
\description{
Linear regression on low-light points to estimate AQY and Rd
}
